% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{batchGSEAplot}
\alias{batchGSEAplot}
\title{Create multiple GSEA plots for different comparisons}
\usage{
batchGSEAplot(
  gsea_list,
  object,
  output_dir = "gsea_plots",
  width = 12,
  height = 8,
  dpi = 300,
  format = "pdf",
  ...
)
}
\arguments{
\item{gsea_list}{named list of GSEAResult objects}

\item{object}{Annot object}

\item{output_dir}{directory to save plots}

\item{width}{plot width}

\item{height}{plot height}

\item{dpi}{resolution}

\item{format}{output format (pdf, png, etc.)}

\item{...}{additional arguments passed to plotGSEA}
}
\description{
Create multiple GSEA plots for different comparisons
}
\examples{
\dontrun{
gsea_results <- list(
  "D7_WT_vs_TFAM" = res1,
  "D14_WT_vs_D7_WT" = res2
)
plots <- batchGSEAplot(gsea_results, object = hsako)
}
}
\author{
Kai Guo
}
