% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeown.R
\name{buildOwn}
\alias{buildOwn}
\title{make annotation database using bioAnno results}
\usage{
buildOwn(dbname, anntype = "GO", OP = NULL, species = "", keytype = "")
}
\arguments{
\item{dbname}{database name from bioAnno}

\item{anntype}{GO or KEGG}

\item{OP}{BP,CC,MF default use all}
}
\description{
make annotation database using bioAnno results
}
\examples{
\dontrun{
fromKEGG(species="ath")
athgo<-makeOwn(dbname="org.ath.eg.db",anntype="GO")
}
}
\author{
Kai Guo
}
