% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResult.R
\name{compareGSEA}
\alias{compareGSEA}
\title{compare GSEA enrichment results across different samples and generate figure}
\usage{
compareGSEA(
  x,
  object,
  gene = NULL,
  pathway = NULL,
  mycol = NULL,
  pvalue = 0.05,
  padj = NULL,
  gseaParam = 1,
  ticksSize = 0.2,
  ncol = 2
)
}
\arguments{
\item{x}{list of GSEAResult}

\item{object}{Annot object}

\item{gene}{list of a vector include all log2FC with gene name (optional)}

\item{pathway}{pathways you want to display (optional)}

\item{mycol}{a vector indicate the colors used for the figure}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{top}{number of terms you want to display,}
}
\description{
compare GSEA enrichment results across different samples and generate figure
}
\examples{
\dontrun{
hsako <- buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
gene1 <- sample(unique(hsako$GeneID),1000)
gene2 <- sample(unique(hsako$GeneID),1000)
fc1<-rnorm(1000,11,2)
names(fc1)<-gene1
fc2<-rnorm(1000,11,2)
names(fc2)<-gene2
resko1 <-richGSEA(fc1,kodata = hsako)
resko2 <-richGSEA(fc2,kodata = hsako)
res<-compareGSEA(list(S1=resko1,S2=resko2),hsako)
}
}
\author{
Kai Guo
}
