% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResult.R
\name{compareResult}
\alias{compareResult}
\title{compare enrichment results across different samples}
\usage{
compareResult(x, pvalue = 0.05, padj = NULL)
}
\arguments{
\item{x}{list of richResults}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}
}
\description{
compare enrichment results across different samples
}
\examples{
\dontrun{
hsako <- buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
hsago <- buildAnnot(species="human",keytype="SYMBOL",anntype = "GO")
gene1 <- sample(unique(hsako$GeneID),1000)
gene2 <- sample(unique(hsako$GeneID),1000)
resko1 <-richKEGG(gene1,kodata = hsako)
resko2 <-richKEGG(gene2,kodata = hsako)
res<-compareResult(list(S1=resko1,S2=resko2))
}
}
\author{
Kai Guo
}
