% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/misc.R
\name{dollar-methods}
\alias{dollar-methods}
\alias{$,Annot-method}
\alias{$,richResult-method}
\alias{$,GSEAResult-method}
\alias{$-Annot}
\alias{$-richResult}
\alias{$-GSEAResult}
\title{S4 dollar accessor methods}
\usage{
\S4method{$}{Annot}(x, name)

\S4method{$}{richResult}(x, name)

\S4method{$}{GSEAResult}(x, name)

\S4method{$}{Annot}(x, name)

\S4method{$}{richResult}(x, name)

\S4method{$}{GSEAResult}(x, name)
}
\arguments{
\item{x}{An S4 object (Annot or richResult)}

\item{name}{column name to extract}
}
\description{
Provide `obj$name` access that extracts from the relevant slot columns.
}
