% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R, R/misc.R
\name{.calculate_Enrichment_Score}
\alias{.calculate_Enrichment_Score}
\title{Calculate enrichment score for a cluster}
\usage{
.calculate_Enrichment_Score(x, df)

.calculate_Enrichment_Score(x, df)
}
\arguments{
\item{x}{character vector of term identifiers}

\item{df}{data.frame with enrichment results containing Pvalue column}
}
\value{
numeric enrichment score
}
\description{
Calculate enrichment score for a cluster

calculate enrichment score
}
\keyword{internal}
