% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.color_scale}
\alias{.color_scale}
\title{Color scale helper (modified from DOSE)}
\usage{
.color_scale(c1 = "pink", c2 = "red")
}
\arguments{
\item{c1}{start color}

\item{c2}{end color}
}
\value{
character vector of 200 colors
}
\description{
Color scale helper (modified from DOSE)
}
\keyword{internal}
