% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggupset.R
\name{.extract_gene_lists}
\alias{.extract_gene_lists}
\title{Extract gene lists from various input types}
\usage{
.extract_gene_lists(x, sep = ",")
}
\arguments{
\item{x}{Named list of character vectors, richResult objects, or GSEAResult objects}

\item{sep}{Separator for gene IDs}
}
\value{
Named list of character vectors
}
\description{
Extract gene lists from various input types
}
\keyword{internal}
