% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_go_dat}
\alias{.get_go_dat}
\title{Get GO annotation terms}
\usage{
.get_go_dat(ont = "BP")
}
\arguments{
\item{ont}{GO ontology: "BP", "CC", or "MF"}
}
\value{
data.frame with rownames as GO IDs and "annotation" column
}
\description{
Get GO annotation terms
}
\keyword{internal}
