% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_kg_dat}
\alias{.get_kg_dat}
\title{Get KEGG pathway annotation terms}
\usage{
.get_kg_dat(builtin = TRUE)
}
\arguments{
\item{builtin}{use built-in KEGG data or fetch from API}
}
\value{
data.frame with rownames as pathway IDs and "annotation" column
}
\description{
Get KEGG pathway annotation terms
}
\keyword{internal}
