% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_kgm.data}
\alias{.get_kgm.data}
\title{Get KEGG module annotation terms}
\usage{
.get_kgm.data()
}
\value{
data.frame module annotation
}
\description{
Get KEGG module annotation terms
}
\keyword{internal}
