% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{.getdb}
\alias{.getdb}
\title{Get OrgDb package name for a species}
\usage{
.getdb(species)
}
\arguments{
\item{species}{common species name}
}
\value{
character OrgDb package name or NULL
}
\description{
Get OrgDb package name for a species
}
\keyword{internal}
