% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{.getdbname}
\alias{.getdbname}
\title{Get OrgDb package name with error handling}
\usage{
.getdbname(species = "human")
}
\arguments{
\item{species}{common species name}
}
\value{
character OrgDb package name (stops on failure)
}
\description{
Get OrgDb package name with error handling
}
\keyword{internal}
