% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getIdx}
\alias{.getIdx}
\title{Get index for color mapping (modified from DOSE)}
\usage{
.getIdx(v, MIN, MAX)
}
\arguments{
\item{v}{numeric value}

\item{MIN}{minimum value}

\item{MAX}{maximum value}
}
\value{
integer index
}
\description{
Get index for color mapping (modified from DOSE)
}
\keyword{internal}
