% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/species.R
\name{.getmsig}
\alias{.getmsig}
\title{msigdb support species}
\usage{
.getmsig(species = "human")

.getmsig(species = "human")
}
\arguments{
\item{species}{common species name}
}
\value{
character scientific name or NULL
}
\description{
msigdb support species

Get msigdbr scientific name
}
\keyword{internal}
