% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{.getrodbname}
\alias{.getrodbname}
\title{Get Reactome species name}
\usage{
.getrodbname(species)
}
\arguments{
\item{species}{common species name}
}
\value{
character Reactome species name or NULL
}
\description{
Get Reactome species name
}
\keyword{internal}
