% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{.getspecies}
\alias{.getspecies}
\alias{.getspeices}
\title{Get KEGG 3-letter species code}
\usage{
.getspecies(species = "human")
}
\arguments{
\item{species}{common species name}
}
\value{
character KEGG code or NULL
}
\description{
Get KEGG 3-letter species code
}
\keyword{internal}
