% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R, R/misc.R
\name{.kappa}
\alias{.kappa}
\title{Compute kappa statistic between two gene sets}
\usage{
.kappa(x, y, geneall)

.kappa(x, y, geneall)
}
\arguments{
\item{x}{comma-separated gene string for first term}

\item{y}{comma-separated gene string for second term}

\item{geneall}{character vector of all genes (background)}
}
\value{
numeric kappa score
}
\description{
Compute kappa statistic between two gene sets

kappa function
}
\keyword{internal}
