% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{.kappa_cluster_internal}
\alias{.kappa_cluster_internal}
\title{calculate kappa cluster}
\usage{
.kappa_cluster_internal(
  x,
  gene = NULL,
  useTerm = FALSE,
  cutoff = 0.5,
  overlap = 0.5,
  minSize = 5,
  escore = 3
)
}
\arguments{
\item{x}{richResult object}

\item{gene}{(Optional).a vector of gene list}

\item{useTerm}{to use the term or not (TRUE/FALSE)}

\item{cutoff}{kappa score threshold for significant dispersion results}

\item{minSize}{minimal number of terms in the cluster}

\item{escore}{kappa enrichment score cutoff value (default: 3)}
}
\description{
calculate kappa cluster
}
\author{
Kai Guo
}
