% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{.make_richResult}
\alias{.make_richResult}
\title{Create a richResult S4 object (shared across enrichment functions)}
\usage{
.make_richResult(
  resultFis,
  detail,
  pvalue,
  padj,
  padj.method,
  input,
  organism,
  ontology,
  keytype,
  sep = ","
)

.make_richResult(
  resultFis,
  detail,
  pvalue,
  padj,
  padj.method,
  input,
  organism,
  ontology,
  keytype,
  sep = ","
)
}
\arguments{
\item{resultFis}{enrichment result data.frame}

\item{detail}{detail data.frame}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{padj.method}{p-value adjustment method}

\item{input}{character vector of input genes}

\item{organism}{organism name}

\item{ontology}{ontology type}

\item{keytype}{key type}

\item{sep}{separator}
}
\value{
richResult S4 object

richResult S4 object
}
\description{
Create a richResult S4 object (shared across enrichment functions)

Create a richResult S4 object (shared across enrichment functions)
}
\keyword{internal}
