% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R, R/misc.R
\name{.merge_term}
\alias{.merge_term}
\title{Merge overlapping term clusters}
\usage{
.merge_term(x, overlap)

.merge_term(x, overlap)
}
\arguments{
\item{x}{named list of term clusters}

\item{overlap}{numeric overlap threshold for merging}
}
\value{
list of merged term clusters
}
\description{
Merge overlapping term clusters

merge term
}
\keyword{internal}
