% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{.species_lookup}
\alias{.species_lookup}
\title{Look up a single column from the species table}
\usage{
.species_lookup(species, column)
}
\arguments{
\item{species}{common species name}

\item{column}{column to retrieve}
}
\value{
character value or NULL
}
\description{
Look up a single column from the species table
}
\keyword{internal}
