% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{.validate_enrichment_input}
\alias{.validate_enrichment_input}
\title{Validate common enrichment inputs}
\usage{
.validate_enrichment_input(
  x,
  annot = NULL,
  pvalue = 0.05,
  padj = NULL,
  padj.method = "BH",
  minSize = 1,
  maxSize = 500
)

.validate_enrichment_input(
  x,
  annot = NULL,
  pvalue = 0.05,
  padj = NULL,
  padj.method = "BH",
  minSize = 1,
  maxSize = 500
)
}
\arguments{
\item{x}{gene vector or data.frame}

\item{annot}{annotation data.frame (2+ columns: GeneID, Term)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff (or NULL)}

\item{padj.method}{p-value adjustment method}

\item{minSize}{minimum significant genes}

\item{maxSize}{maximum significant genes}
}
\description{
Validate common enrichment inputs

Validate common enrichment inputs
}
\keyword{internal}
