% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validateParams}
\alias{.validateParams}
\title{Validate enrichment parameters}
\usage{
.validateParams(
  pvalue = 0.05,
  padj = NULL,
  minSize = 2,
  maxSize = 500,
  minGSSize = 10,
  maxGSSize = 500,
  func_name = "enrich"
)
}
\arguments{
\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{minSize}{minimum gene set size}

\item{maxSize}{maximum gene set size}

\item{minGSSize}{minimum gene set size for annotation}

\item{maxGSSize}{maximum gene set size for annotation}

\item{func_name}{name of calling function}
}
\description{
Validate enrichment parameters
}
\keyword{internal}
