% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich,ANY,Annot-method}
\alias{enrich,ANY,Annot-method}
\title{Enrichment analysis function}
\usage{
\S4method{enrich}{ANY,Annot}(
  x,
  object,
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  keytype = "",
  filename = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  padj.method = "BH",
  builtin = TRUE,
  sep = ","
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{object}{annotation data}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{keytype}{keytype for input genes}

\item{filename}{output filename}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}

\item{ontology}{ontology type}

\item{bulitin}{use KEGG bulit in KEGG annotation or not(set FALSE if you want use newest KEGG data)}
}
\description{
Enrichment analysis function
}
\author{
Kai Guo
}
