% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich,ANY,data.frame-method}
\alias{enrich,ANY,data.frame-method}
\title{Enrichment analysis function}
\usage{
\S4method{enrich}{ANY,data.frame}(
  x,
  object,
  ontology = "",
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  keytype = "",
  filename = NULL,
  minSize = 2,
  maxSize = 500,
  minGSSize = 10,
  maxGSSize = 500,
  keepRich = TRUE,
  padj.method = "BH",
  sep = ","
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{object}{annotation data}

\item{ontology}{ontology type}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{keytype}{keytype for input genes}

\item{filename}{output filename}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{minGSSize}{minimal size of genes annotated by ontology term for testing.}

\item{maxGSSize}{maximal size of each geneset for analyzing}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
Enrichment analysis function
}
\examples{
\dontrun{
  hsago <- buildAnnot(species="human",keytype="SYMBOL",anntype = "GO")
  hsago <- as.data.frame(hsago)
  gene <- sample(unique(hsago$GeneID),1000)
  res<-enrich(gene,godata = hsago)
}
}
\author{
Kai Guo
}
