% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{enrich}
\alias{enrich}
\title{enrich method}
\usage{
enrich(
  x,
  annot,
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  ontology = "",
  keytype = "",
  filename = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  padj.method = "BH",
  sep = ",",
  ...
)
}
\arguments{
\item{x}{vector contains gene names}

\item{annot}{Annotation object or dataframe}

\item{...}{additional parameters}
}
\value{
richResult
}
\description{
enrich
}
\examples{
\dontrun{
library(bioAnno)
fromKEGG(species="ath")
athgo<-buildOwn(dbname="org.ath.eg.db",anntype="GO")
gene=sample(unique(athgo$GeneID),1000)
res<-enrich(gene,athgo)
}
}
\author{
Kai Guo
}
