% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich_internal}
\alias{enrich_internal}
\title{Enrichment analysis for any type of annotation data}
\usage{
enrich_internal(
  x,
  object,
  ontology = "",
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  minSize = 1,
  maxSize = 500,
  keepRich = TRUE,
  keytype = "",
  filename = NULL,
  padj.method = "BH",
  sep = ","
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{object}{annotation data}

\item{ontology}{ontology type}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{keytype}{keytype for input genes}

\item{filename}{output filename}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
Enrichment analysis for any type of annotation data
}
\author{
Kai Guo
}
