% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{filterPathways}
\alias{filterPathways}
\title{Filter pathways by multiple criteria}
\usage{
filterPathways(
  gseaRes,
  nes_cutoff = NULL,
  padj_cutoff = 0.05,
  pval_cutoff = 0.05,
  min_size = NULL,
  max_size = NULL,
  direction = "both"
)
}
\arguments{
\item{gseaRes}{GSEAResult object}

\item{nes_cutoff}{absolute NES cutoff}

\item{padj_cutoff}{padj cutoff}

\item{pval_cutoff}{pval cutoff (used if padj_cutoff is NULL)}

\item{min_size}{minimum pathway size}

\item{max_size}{maximum pathway size}

\item{direction}{filter by direction ("up", "down", or "both")}
}
\description{
Filter pathways by multiple criteria
}
\examples{
\dontrun{
filtered <- filterPathways(gseaRes, nes_cutoff = 1.5, padj_cutoff = 0.01, direction = "up")
}
}
\author{
Kai Guo
}
