% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{getann}
\alias{getann}
\title{build annotation for kegg}
\usage{
getann(ontype = "GO", builtin = TRUE)

getann(ontype = "GO", builtin = TRUE)
}
\arguments{
\item{ontype}{one of "GO", "KEGG", "Module"}

\item{builtin}{use built-in annotation or not (default TRUE)}
}
\value{
data.frame with annotation terms
}
\description{
build annotation for kegg

Build annotation lookup for GO, KEGG, or Module
}
\examples{
\dontrun{
annot = getann("GO")
}
}
\author{
Kai Guo
}
