% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{getPathways}
\alias{getPathways}
\title{Get available pathways from GSEA results}
\usage{
getPathways(
  gseaRes,
  padj_cutoff = 0.05,
  pval_cutoff = 0.05,
  min_size = NULL,
  max_size = NULL
)
}
\arguments{
\item{gseaRes}{GSEAResult object}

\item{padj_cutoff}{padj cutoff for filtering (if NULL, no filtering applied)}

\item{pval_cutoff}{pvalue cutoff for filtering (used if padj_cutoff is NULL)}

\item{min_size}{minimum pathway size}

\item{max_size}{maximum pathway size}
}
\description{
Get available pathways from GSEA results
}
\examples{
\dontrun{
pathways <- getPathways(gseaRes, padj_cutoff = 0.05)
}
}
\author{
Kai Guo
}
