% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{getPathwayStats}
\alias{getPathwayStats}
\title{Get pathway statistics summary}
\usage{
getPathwayStats(gseaRes, top = 10)
}
\arguments{
\item{gseaRes}{GSEAResult object}

\item{top}{number of top pathways to summarize}
}
\description{
Get pathway statistics summary
}
\examples{
\dontrun{
summary <- getPathwayStats(gseaRes, top = 20)
print(summary)
}
}
\author{
Kai Guo
}
