% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.R
\name{ggbar,richResult-method}
\alias{ggbar,richResult-method}
\alias{ggbar,data.frame-method}
\title{barplot for Enrichment results}
\usage{
\S4method{ggbar}{richResult}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  fontsize.text = 3,
  angle = 75,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{ggbar}{data.frame}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  fontsize.text = 3,
  angle = 75,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{dataframe of enrichment results}

\item{top}{number of terms you want to display,}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{angle}{label angle of xlim}

\item{filename}{figure output name}

\item{font.size}{font size for xlim or ylim}
}
\description{
barplot for Enrichment results

barplot for Enrichment result
}
\author{
Kai Guo
}
