% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new.R
\name{ggcircbar_internal}
\alias{ggcircbar_internal}
\title{Circular barplot (polar bar) of enrichment results}
\usage{
ggcircbar_internal(
  resultFis,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{data.frame of enrichment results}

\item{top}{number of top terms to display (default 15)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted-p cutoff}

\item{usePadj}{colour by adjusted p-value?}

\item{low}{low-end gradient colour}

\item{high}{high-end gradient colour}

\item{short}{shorten long term names}

\item{filename}{if non-NULL, save figure}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
Circular barplot (polar bar) of enrichment results
}
\keyword{internal}
