% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggdot}
\alias{ggdot}
\title{ggdot method}
\usage{
ggdot(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  padj = NULL,
  usePadj = TRUE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult or dataframe}

\item{...}{additional parameters}
}
\value{
plot
}
\description{
ggdot
}
\examples{
\dontrun{
  hsago<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsago$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggdot(res)
}
}
\author{
Kai Guo
}
