% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdot.R
\name{ggdot,richResult-method}
\alias{ggdot,richResult-method}
\alias{ggdot,data.frame-method}
\title{dotplot for Enrichment results}
\usage{
\S4method{ggdot}{richResult}(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  padj = NULL,
  usePadj = TRUE,
  filename = NULL,
  width = 10,
  height = 8
)

\S4method{ggdot}{data.frame}(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  padj = NULL,
  usePadj = TRUE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{object}{data.frame object}

\item{top}{number of terms you want to display,}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{filename}{figure output name}

\item{font.size}{font size for xlim or ylim}
}
\description{
dotplot for Enrichment results

dotplot for Enrichment results
}
\author{
Kai Guo

Kai Guo
}
