% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new.R
\name{ggecdf_internal}
\alias{ggecdf_internal}
\title{ECDF step plot of GSEA ranked gene sets}
\usage{
ggecdf_internal(
  resultFis,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{data.frame with GSEA results (must contain NES column)}

\item{top}{number of top terms to display (default 30)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted-p cutoff}

\item{line.color}{line colour (default \code{"#1b7837"})}

\item{line.size}{line size (default 1.2)}

\item{point.size}{point size (default 3)}

\item{short}{shorten long term names}

\item{filename}{if non-NULL, save figure}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
ECDF step plot of GSEA ranked gene sets
}
\keyword{internal}
