% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new.R
\name{gggenebar_internal}
\alias{gggenebar_internal}
\title{Barplot with gene names embedded inside bars}
\usage{
gggenebar_internal(
  resultFis,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{data.frame of enrichment results}

\item{top}{number of top terms (default 10)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted-p cutoff}

\item{max.genes}{max gene names shown inside each bar (default 5)}

\item{bar.color}{bar colour (default \code{"#b2182b"})}

\item{bar.alpha}{bar transparency (default 0.85)}

\item{text.size}{size of gene labels inside bars (default 3)}

\item{short}{shorten long term names}

\item{sep}{separator for gene IDs}

\item{filename}{if non-NULL, save figure}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
Barplot with gene names embedded inside bars
}
\keyword{internal}
