% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{ggGSEA}
\alias{ggGSEA}
\title{plot the gsea result}
\usage{
ggGSEA(term, object, gseaRes, top = 10, default = TRUE)
}
\arguments{
\item{term}{the significant term}

\item{object}{Annot object}

\item{gseaRes}{GSEAResult object}

\item{top}{number of top pathways to display}

\item{default}{whether to use default table plot or individual enrichment plots}
}
\description{
plot the gsea result
}
\examples{
\dontrun{
set.seed(123)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
ggGSEA(term = res$pathway, object = hsako, gseaRes = res, default = FALSE)
}
}
\author{
Kai Guo
}
