% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Plot compare heatmap of Enrichment result among DEG groups}
\usage{
ggheatmap(
  richRes,
  top = 50,
  colnames = NULL,
  xsize = 6,
  ysize = 6,
  usePadj = FALSE,
  horizontal = FALSE,
  returnData = FALSE,
  ...
)
}
\arguments{
\item{richRes}{list of enrichment object}

\item{top}{the number of Terms you want to display}

\item{colnames}{the compare DEG group names}

\item{xsize}{size of group name}

\item{ysize}{size of Terms name}
}
\description{
Plot compare heatmap of Enrichment result among DEG groups
}
\author{
Kai Guo
}
