% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new.R
\name{ggNES_internal}
\alias{ggNES_internal}
\title{Bidirectional NES barplot of GSEA results}
\usage{
ggNES_internal(
  resultFis,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{data.frame with GSEA results (must contain NES column)}

\item{top}{number of top terms to display (default 20)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted-p cutoff}

\item{up.color}{colour for positive NES (default \code{"#b2182b"})}

\item{down.color}{colour for negative NES (default \code{"#2166ac"})}

\item{short}{shorten long term names}

\item{filename}{if non-NULL, save figure}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
Bidirectional NES barplot of GSEA results
}
\keyword{internal}
