% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggnetplot}
\alias{ggnetplot}
\title{ggnetplot method}
\usage{
ggnetplot(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  useTerm = TRUE,
  low = "orange",
  high = "red",
  writeCyt = FALSE,
  cytoscapeFile = "network-file-for-cytoscape.txt",
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = layout.fruchterman.reingold,
  savefig = FALSE,
  filename = "network",
  width = 7,
  height = 7,
  node.alpha = 0.7,
  repel = TRUE,
  segment.size = 0.2,
  sep = ",",
  ...
)
}
\arguments{
\item{object}{richResult or dataframe}

\item{...}{additional parameters}
}
\value{
plot
}
\description{
ggnetplot
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggnetplot(res)
}
}
\author{
Kai Guo
}
