% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggnetwork}
\alias{ggnetwork}
\title{ggnetwork method}
\usage{
ggnetwork(
  object,
  gene,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "orange",
  high = "red",
  weightcut = 0.2,
  useTerm = TRUE,
  writeCyt = FALSE,
  cytoscapeFile = "network-file-for-cytoscape.txt",
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = layout.fruchterman.reingold,
  savefig = FALSE,
  visNet = FALSE,
  smooth = TRUE,
  nodeselect = FALSE,
  edit = FALSE,
  savehtml = FALSE,
  filename = "network",
  width = 7,
  height = 7,
  segment.size = 0.2,
  node.alpha = 0.7,
  sep = ",",
  ...
)
}
\arguments{
\item{object}{richResult object or dataframe}

\item{...}{additional paramters}
}
\value{
plot
}
\description{
ggnetwork
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggnetwork(res)
}
}
