% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new.R
\name{ggscatter_internal}
\alias{ggscatter_internal}
\title{Labeled scatter plot of enrichment results}
\usage{
ggscatter_internal(
  resultFis,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 3,
  label.size = 3,
  label.top = 5,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{data.frame of enrichment results}

\item{top}{number of top terms to display (default 20)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted-p cutoff}

\item{usePadj}{colour by adjusted p-value?}

\item{low}{low-end gradient colour}

\item{high}{high-end gradient colour}

\item{point.size}{point size (default 5)}

\item{label.size}{text label size (default 3)}

\item{short}{shorten long term names}

\item{filename}{if non-NULL, save figure}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
Labeled scatter plot of enrichment results
}
\keyword{internal}
