% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{GO_child}
\alias{GO_child}
\title{Get all children terms of node}
\usage{
GO_child(node = "GO:0008150", ontology = "BP")

GO_child(node = "GO:0008150", ontology = "BP")
}
\arguments{
\item{node}{GO term ID (default: "GO:0008150" for BP root)}

\item{ontology}{one of "BP", "CC", "MF"}
}
\value{
character vector of child GO IDs
}
\description{
Get all children terms of node

Get all children terms of a GO node
}
\author{
Kai Guo
}
