% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{godata}
\alias{godata}
\title{GO annotation data for human}
\format{
A data.frame with columns:
\describe{
  \item{GeneID}{Gene identifier}
  \item{GO}{GO term identifier}
  \item{Ontology}{GO ontology (BP, CC, MF)}
}
}
\source{
Generated from Bioconductor org.Hs.eg.db and GO.db
}
\usage{
godata
}
\description{
A data.frame containing GO annotation mappings for human genes.
}
\keyword{datasets}
