% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{GSEAResult-class}
\alias{GSEAResult-class}
\alias{show,GSEAResult-method}
\alias{summary,GSEAResult-method}
\alias{plot,GSEAResult-method}
\title{Class "GSEAResult"
This class represents the result from GSEA analysis}
\description{
Class "GSEAResult"
This class represents the result from GSEA analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{enrichment analysis results}

\item{\code{pvalueCutoff}}{cutoff pvalue}

\item{\code{pAdjustMethod}}{pvalue adjust method}

\item{\code{padjCutoff}}{pvalue adjust cutoff value}

\item{\code{genenumber}}{number of input genes}

\item{\code{organism}}{organism used}

\item{\code{ontology}}{biological ontology}

\item{\code{gene}}{Gene IDs}

\item{\code{keytype}}{Gene ID type}

\item{\code{sep}}{character string used to separate the genes when concatenating}
}}

\author{
Kai Guo
}
\keyword{classes}
