% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/species.R
\name{idconvert}
\alias{idconvert}
\title{Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package}
\usage{
idconvert(species, keys, fkeytype, tkeytype)

idconvert(species, keys, fkeytype, tkeytype)
}
\arguments{
\item{species}{species common name (see showData())}

\item{keys}{character vector of gene IDs to convert}

\item{fkeytype}{source key type (e.g. "SYMBOL", "ENTREZID")}

\item{tkeytype}{target key type (e.g. "ENTREZID", "SYMBOL")}
}
\value{
named character vector of converted IDs
}
\description{
Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package

Convert gene IDs between types using Bioconductor annotation packages
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  gene=sample(unique(hsako$GeneID),1000)
  id<-idconvert(species="human",keys=gene,fkeytype="SYMBOL",tkeytype="ENTREZID")
}
\dontrun{
  id <- idconvert(species="human", keys=c("TP53","BRCA1"),
                  fkeytype="SYMBOL", tkeytype="ENTREZID")
}
}
\author{
Kai Guo
}
