% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg.db}
\alias{kegg.db}
\title{KEGG pathway database}
\format{
A data.frame with KEGG pathway identifiers and descriptions
}
\source{
KEGG database via KEGGREST
}
\usage{
kegg.db
}
\description{
A data.frame containing KEGG pathway and species mapping information.
}
\keyword{datasets}
