% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module}
\alias{module}
\title{KEGG module data}
\format{
A data.frame with KEGG module identifiers and descriptions
}
\source{
KEGG database via KEGGREST
}
\usage{
module
}
\description{
A data.frame containing KEGG module information.
}
\keyword{datasets}
