% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{overlap}
\alias{overlap}
\title{Compute Jaccard overlap between two vectors}
\usage{
overlap(x, y)

overlap(x, y)
}
\arguments{
\item{x}{first vector or list}

\item{y}{second vector or list}
}
\value{
numeric Jaccard index
}
\description{
Compute Jaccard overlap between two vectors

Jaccard overlap between two vectors
}
\keyword{internal}
