% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{parGSEA}
\alias{parGSEA}
\title{GSEA Enrichment analysis function for data frame with log2FC and name specifc}
\usage{
parGSEA(
  x,
  object,
  log2FC = "log2FoldChange",
  gene.col = NULL,
  simplify = FALSE,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  KEGG = FALSE,
  minSize = 15,
  ontology = "",
  maxSize = 500,
  padj.method = "BH",
  organism = NULL,
  table = TRUE,
  sep = ","
)
}
\arguments{
\item{x}{data.frame include the gene name and log2FC}

\item{object}{annotation file for all genes}

\item{log2FC}{the column name for log2FoldChange}

\item{gene.col}{the gene name column if rownames won't be used as gene name}

\item{simplify}{boolean value to indicate if the simple table shoule be returned}

\item{pvalue}{pvalue cutoff value}

\item{padj}{adjust p value cut off method}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{padj.method}{p value adjust method}

\item{table}{leadingEdge as vector}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
GSEA Enrichment analysis function for data frame with log2FC and name specifc
}
\examples{
\dontrun{
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
hsako<-as.data.frame(hsako)
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
d<-data.frame(Gene=name,log2FoldChange=gene)
rownames(d)<-d$Gene
res<-parGSEA(d,object = hsako)
}
}
\author{
Kai Guo
}
