% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path}
\alias{path}
\title{KEGG pathway data}
\format{
A data.frame with KEGG pathway identifiers and descriptions
}
\source{
KEGG database via KEGGREST
}
\usage{
path
}
\description{
A data.frame containing KEGG pathway information.
}
\keyword{datasets}
