% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway}
\alias{pathway}
\title{KEGG pathway hierarchy}
\format{
A data.frame with columns for pathway identifiers and hierarchy levels
}
\source{
KEGG database
}
\usage{
pathway
}
\description{
A data.frame containing KEGG pathway hierarchy with Level1 and Level2 categories.
}
\keyword{datasets}
