% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/misc.R
\name{rbind.GSEAResult}
\alias{rbind.GSEAResult}
\title{rbind for GSEAResult objects}
\usage{
\method{rbind}{GSEAResult}(...)

\method{rbind}{GSEAResult}(...)
}
\arguments{
\item{...}{GSEAResult objects to combine}
}
\description{
rbind for GSEAResult objects

rbind generic function for GSEAResult object
}
\author{
Kai Guo
}
