% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/misc.R
\name{rbind.richResult}
\alias{rbind.richResult}
\title{rbind for richResult objects}
\usage{
\method{rbind}{richResult}(...)

\method{rbind}{richResult}(...)
}
\arguments{
\item{...}{richResult objects to combine}
}
\description{
rbind for richResult objects

rbind generic function for richResult object
}
\author{
Kai Guo
}
