% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readGMT}
\alias{readGMT}
\title{Read gene sets from a GMT file}
\usage{
readGMT(file, species = "custom", keytype = "SYMBOL")
}
\arguments{
\item{file}{path to a GMT file}

\item{species}{species name (default: "custom")}

\item{keytype}{gene ID type in the file (default: "SYMBOL")}
}
\value{
An \code{Annot} object
}
\description{
Reads a Gene Matrix Transposed (GMT) file and returns an Annot object
for use with \code{enrich()} or \code{richGSEA()}.
}
\details{
GMT format: each line is tab-separated with:
\enumerate{
  \item Gene set name
  \item Description (or URL)
  \item Gene 1, Gene 2, ... (tab-separated)
}
}
\examples{
\dontrun{
# Read MSigDB hallmark gene sets
annot <- readGMT("h.all.v2023.2.Hs.symbols.gmt")
res <- enrich(my_genes, annot)

# Read with species info
annot <- readGMT("my_pathways.gmt", species = "human", keytype = "SYMBOL")
}
}
\author{
Junguk Hur
}
