% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/misc.R
\name{result_detail_methods}
\alias{result_detail_methods}
\alias{result,richResult-method}
\alias{result,GSEAResult-method}
\alias{detail,richResult-method}
\alias{detail,GSEAResult-method}
\title{S4 methods for result() and detail()}
\usage{
\S4method{result}{richResult}(x)

\S4method{result}{GSEAResult}(x)

\S4method{detail}{richResult}(x)

\S4method{detail}{GSEAResult}(x)

\S4method{result}{richResult}(x)

\S4method{result}{GSEAResult}(x)

\S4method{detail}{richResult}(x)

\S4method{detail}{GSEAResult}(x)
}
\arguments{
\item{x}{A \code{richResult} or \code{GSEAResult} object}
}
\value{
data.frame

A data frame
}
\description{
These methods return \code{as.data.frame(x@result)} or \code{x@detail},
depending on the class, effectively exposing object contents.
}
