% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zzz_ggplots_aliases.R
\name{richBar}
\alias{richBar}
\alias{ggbar}
\title{ggbar method}
\usage{
richBar(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  fontsize.text = 3,
  angle = 0,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  horiz = TRUE,
  ...
)

richBar(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  fontsize.text = 3,
  angle = 0,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  horiz = TRUE,
  ...
)
}
\arguments{
\item{object}{enrichment result or data.frame}

\item{top}{number of terms you want to display}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{order}{order by Term or richFactor}

\item{usePadj}{use p adjust value as color or not}

\item{fontsize.x}{fontsize of x axis}

\item{fontsize.y}{fontsize of y axis}

\item{short}{automatic short name or not}

\item{fontsize.text}{font size for bar text labels}

\item{angle}{angle of x axis text}

\item{orderp}{order by p value (adjusted p value)}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{horiz}{horizontal bar plot or not}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
richBar
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  richBar(res)
}
}
\author{
Kai Guo
}
