% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richCircle}
\alias{richCircle}
\alias{richCircle,richResult-method}
\alias{richCircle,data.frame-method}
\alias{richCircle,GSEAResult-method}
\alias{ggcircbar}
\alias{ggCircBar}
\title{Circular barplot of enrichment results}
\usage{
richCircle(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richCircle}{richResult}(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richCircle}{data.frame}(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richCircle}{GSEAResult}(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richCircle(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richCircle(
  object,
  top = 15,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult, data.frame, or GSEAResult}

\item{top}{number of terms (default 15)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{usePadj}{colour by adjusted p-value?}

\item{low}{low-end gradient colour}

\item{high}{high-end gradient colour}

\item{short}{shorten term names}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Circular barplot of enrichment results
}
\examples{
\dontrun{
  res <- richKEGG(gene, kodata = hsako)
  richCircle(res)
}
}
