% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ggdot.R,
%   R/zzz_ggplots_aliases.R
\name{richDot}
\alias{richDot}
\alias{richDot,richResult-method}
\alias{richDot,data.frame-method}
\alias{ggdot}
\title{richDot method}
\usage{
richDot(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  padj = NULL,
  usePadj = TRUE,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richDot}{richResult}(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  padj = NULL,
  usePadj = TRUE,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)

\S4method{richDot}{data.frame}(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  padj = NULL,
  usePadj = TRUE,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)

richDot(
  object,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  low = "lightpink",
  high = "red",
  alpha = 0.7,
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  padj = NULL,
  usePadj = TRUE,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{data.frame of enrichment results}

\item{top}{number of terms you want to display}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{order}{order by Term or richFactor}

\item{low}{low color}

\item{high}{high color}

\item{alpha}{transparency alpha}

\item{font.x}{font of x axis}

\item{font.y}{font of y axis}

\item{fontsize.x}{fontsize of x axis}

\item{fontsize.y}{fontsize of y axis}

\item{short}{automatic short name or not}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{orderp}{order by p value (adjusted p value)}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
richDot
}
\examples{
\dontrun{
  hsako <- buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene <- sample(unique(hsako$GeneID),1000)
  res <- richKEGG(gene,kodata = hsako)
  richDot(res)
}
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  richDot(res)
}
\dontrun{
  hsago<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsago$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  richDot(result(res))
}
}
\author{
Kai Guo
}
