% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richECDF}
\alias{richECDF}
\alias{richECDF,GSEAResult-method}
\alias{richECDF,data.frame-method}
\alias{ggecdf}
\alias{ggECDF}
\title{ECDF step plot of GSEA ranked gene sets}
\usage{
richECDF(
  object,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richECDF}{GSEAResult}(
  object,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richECDF}{data.frame}(
  object,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richECDF(
  object,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richECDF(
  object,
  top = 30,
  pvalue = 0.05,
  padj = NULL,
  line.color = "#1b7837",
  line.size = 1.2,
  point.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{GSEAResult or data.frame with NES column}

\item{top}{number of terms (default 30)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{line.color}{line colour}

\item{line.size}{line size}

\item{point.size}{point size}

\item{short}{shorten term names}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
ECDF step plot of GSEA ranked gene sets
}
\examples{
\dontrun{
  res <- richGSEA(genelist, gseadata)
  richECDF(res)
}
}
