% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richGeneBar}
\alias{richGeneBar}
\alias{richGeneBar,richResult-method}
\alias{richGeneBar,data.frame-method}
\alias{richGeneBar,GSEAResult-method}
\alias{gggenebar}
\alias{ggGeneBar}
\title{Barplot with gene names embedded inside bars}
\usage{
richGeneBar(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richGeneBar}{richResult}(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richGeneBar}{data.frame}(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richGeneBar}{GSEAResult}(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richGeneBar(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richGeneBar(
  object,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "lightpink",
  high = "red",
  max.genes = 5,
  bar.color = "#b2182b",
  bar.alpha = 0.85,
  text.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult, data.frame, or GSEAResult}

\item{top}{number of terms (default 10)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{usePadj}{logical; use adjusted p-value for colour gradient? (default TRUE)}

\item{low}{low-end gradient colour (default \code{"lightpink"})}

\item{high}{high-end gradient colour (default \code{"red"})}

\item{max.genes}{max gene names inside each bar}

\item{bar.color}{bar colour (unused; bars coloured by significance)}

\item{bar.alpha}{bar transparency}

\item{text.size}{gene label size}

\item{short}{shorten term names}

\item{sep}{gene-ID separator}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Barplot with gene names embedded inside bars
}
\examples{
\dontrun{
  res <- richKEGG(gene, kodata = hsako)
  richGeneBar(res)
}
}
