% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richGeneDot}
\alias{richGeneDot}
\alias{richGeneDot,richResult-method}
\alias{richGeneDot,data.frame-method}
\alias{richGeneDot,GSEAResult-method}
\alias{gggenedot}
\alias{ggGeneDot}
\title{Gene-pathway association dotplot}
\usage{
richGeneDot(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)

\S4method{richGeneDot}{richResult}(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)

\S4method{richGeneDot}{data.frame}(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)

\S4method{richGeneDot}{GSEAResult}(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)

richGeneDot(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)

richGeneDot(
  object,
  fc = NULL,
  top = 10,
  max.genes = 30,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#2166ac",
  mid = "white",
  high = "#b2182b",
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 12,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult, data.frame, or GSEAResult}

\item{fc}{named numeric vector of fold-changes (names = gene IDs). When NULL, colour by -log10(Pvalue/Padj)}

\item{top}{number of terms (default 10)}

\item{max.genes}{max genes per term (default 30)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{usePadj}{logical; use adjusted p-value for colour gradient? (default TRUE)}

\item{low}{colour for negative FC (or low significance when fc=NULL)}

\item{mid}{colour for zero FC}

\item{high}{colour for positive FC (or high significance when fc=NULL)}

\item{short}{shorten term names}

\item{sep}{gene-ID separator}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Gene-pathway association dotplot
}
\examples{
\dontrun{
  res <- richKEGG(gene, kodata = hsako)
  richGeneDot(res, fc = my_foldchanges)
  richGeneDot(res)  # no fc: colour by significance
}
}
