% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richGO}
\alias{richGO}
\title{richGO method}
\usage{
richGO(
  x,
  godata,
  ontology = "BP",
  pvalue = 0.05,
  padj = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  filename = NULL,
  padj.method = "BH",
  ...
)
}
\arguments{
\item{x}{vector contains gene names}

\item{godata}{Annotation object or dataframe}

\item{...}{additional parameters}
}
\value{
richResult
}
\description{
richGO
}
\author{
Kai Guo
}
