% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{richGSEA,ANY,Annot-method}
\alias{richGSEA,ANY,Annot-method}
\title{GSEA Enrichment analysis function}
\usage{
\S4method{richGSEA}{ANY,Annot}(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  KEGG = FALSE,
  minSize = 15,
  ontology = "",
  maxSize = 500,
  padj.method = "BH",
  organism = NULL,
  table = TRUE,
  sep = ","
)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{object}{Annot object}

\item{keytype}{gene ID type}

\item{pvalue}{pvalue cutoff value}

\item{padj}{adjust p value cut off}

\item{KEGG}{logical, whether KEGG GSEA were performed}

\item{minSize}{Minimal size of a gene set to test}

\item{ontology}{ontology type}

\item{maxSize}{Maximal size of a gene set to test}

\item{padj.method}{pvalue adjust method}

\item{organism}{organism name}

\item{table}{leadingEdge as vector}

\item{sep}{character string used to separate the genes}
}
\description{
GSEA Enrichment analysis function
}
\examples{
\dontrun{
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
}
}
\author{
Kai Guo
}
