% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{richGSEA,ANY,data.frame-method}
\alias{richGSEA,ANY,data.frame-method}
\title{GSEA Enrichment analysis function}
\usage{
\S4method{richGSEA}{ANY,data.frame}(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  minSize = 15,
  ontology = ontology,
  maxSize = 500,
  nperm = 5000,
  filename = NULL,
  padj.method = "BH",
  organism = NULL,
  table = TRUE,
  sep = ","
)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{object}{annotation file for all genes}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{filename}{output filename}

\item{padj.method}{p value adjust method}

\item{table}{leadingEdge as vector}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
GSEA Enrichment analysis function
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  name=sample(unique(hsako$GeneID),1000)
  gene<-rnorm(1000)
  names(gene)<-name
  res<-richKEGG(gene, hsako)
}
}
\author{
Kai Guo
}
