% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richGSEA}
\alias{richGSEA}
\title{richGSEA method}
\usage{
richGSEA(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  minSize = 15,
  ontology = "",
  maxSize = 500,
  nperm = 5000,
  filename = NULL,
  padj.method = "BH",
  organism = NULL,
  ...
)
}
\arguments{
\item{x}{vector contains gene names}

\item{...}{additional parameters}

\item{Annot}{object or dataframe}
}
\value{
GSEAResult
}
\description{
richGSEA
}
\author{
Kai Guo
}
