% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R, R/zzz_ggplots_aliases.R
\name{richHeatmap}
\alias{richHeatmap}
\alias{ggheatmap}
\title{Plot compare heatmap of Enrichment result among DEG groups}
\usage{
richHeatmap(
  richRes,
  top = 50,
  colnames = NULL,
  xsize = 6,
  ysize = 6,
  usePadj = FALSE,
  horizontal = FALSE,
  returnData = FALSE,
  ...
)

ggheatmap(
  richRes,
  top = 50,
  colnames = NULL,
  xsize = 6,
  ysize = 6,
  usePadj = FALSE,
  horizontal = FALSE,
  returnData = FALSE,
  ...
)
}
\arguments{
\item{richRes}{list of enrichment object}

\item{top}{the number of Terms you want to display}

\item{colnames}{the compare DEG group names}

\item{xsize}{size of group name}

\item{ysize}{size of Terms name}

\item{usePadj}{use adjusted p-value instead of p-value}

\item{horizontal}{flip the heatmap horizontally}

\item{returnData}{return the data instead of the plot}

\item{...}{additional arguments}
}
\description{
Plot compare heatmap of Enrichment result among DEG groups
}
\examples{
\dontrun{
  hsago <- buildAnnot(species="human", keytype="SYMBOL", anntype="GO")
  hsako <- buildAnnot(species="human", keytype="SYMBOL", anntype="KEGG")
  gene <- sample(unique(hsago$GeneID), 1000)
  resgo <- richGO(gene, godata = hsago, ontology = "BP")
  resko <- richKEGG(gene, kodata = hsako)
  richHeatmap(list(GO = resgo, KEGG = resko), top = 50)
}
}
\author{
Kai Guo
}
