% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richLevel.R
\name{richLevel,ANY,data.frame-method}
\alias{richLevel,ANY,data.frame-method}
\title{Pathway Enrichment analysis for different level function}
\usage{
\S4method{richLevel}{ANY,data.frame}(
  x,
  kodata,
  level = "Level2",
  pvalue = 0.05,
  padj = NULL,
  organism = NULL,
  keytype = "SYMBOL",
  ontology = "KEGG",
  minSize = 2,
  maxSize = 500,
  minGSSize = 10,
  maxGSSize = 500,
  keepRich = TRUE,
  filename = NULL,
  padj.method = "BH",
  sep = ","
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{kodata}{KEGG annotation data}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{organism}{organism}

\item{keytype}{keytype for input genes}

\item{ontology}{ontology type}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{minGSSize}{minimal size of genes annotated by ontology term for testing.}

\item{maxGSSize}{maximal size of each geneset for analyzing}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}
}
\description{
Pathway Enrichment analysis for different level function
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  hsako<-as.data.frame(hsako)
  gene=sample(unique(hsako$GeneID),1000)
  res<-richLevel(gene,kodata = hsako,level="Level2")
}
}
\author{
Kai Guo
}
