% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richNES}
\alias{richNES}
\alias{richNES,GSEAResult-method}
\alias{richNES,data.frame-method}
\alias{ggNES}
\alias{ggnes}
\title{Bidirectional NES barplot of GSEA results}
\usage{
richNES(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richNES}{GSEAResult}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richNES}{data.frame}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richNES(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richNES(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  up.color = "#b2182b",
  down.color = "#2166ac",
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{GSEAResult or data.frame with NES column}

\item{top}{number of terms (default 20)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{up.color}{colour for positive NES}

\item{down.color}{colour for negative NES}

\item{short}{shorten term names}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Bidirectional NES barplot of GSEA results
}
\examples{
\dontrun{
  res <- richGSEA(genelist, gseadata)
  richNES(res)
}
}
