% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork.R
\name{ggnetwork_internal}
\alias{ggnetwork_internal}
\title{generate network based on Enrichment results}
\usage{
ggnetwork_internal(
  object = object,
  gene = gene,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "orange",
  high = "red",
  weightcut = 0.2,
  useTerm = TRUE,
  writeCyt = FALSE,
  cytoscapeFile = NULL,
  cytoscapeFormat = "graphml",
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = "fruchtermanreingold",
  savefig = FALSE,
  visNet = FALSE,
  smooth = TRUE,
  nodeselect = FALSE,
  edit = FALSE,
  savehtml = FALSE,
  filename = NULL,
  width = 7,
  height = 7,
  segment.size = 0.2,
  node.alpha = 0.7,
  sep = ",",
  ...
)
}
\arguments{
\item{object}{richResult,GSEAResult object or dataframe}

\item{gene}{vector contains gene names or dataframe with DEGs information}

\item{top}{number of terms to display}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{weightcut}{cutoff value for edge weight}

\item{useTerm}{use Term name instead of Annotation ID}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{cytoscapeFormat}{Character string giving the output file format}

\item{label.color}{label color}

\item{label.size}{label size}

\item{node.shape}{shape of the nodes}

\item{layout}{layout method ('fruchtermanreingold','kamadakawai','target','circle')}

\item{savefig}{save the figure or not}

\item{visNet}{use visNetwork for interactive plot}

\item{smooth}{smooth edges in visNetwork}

\item{nodeselect}{enable node selection in visNetwork}

\item{edit}{enable editing in visNetwork}

\item{savehtml}{save visNetwork as HTML file}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{segment.size}{size for label segment}

\item{node.alpha}{alpha-transparency scales}

\item{sep}{character string used to separate the genes when concatenating}

\item{...}{additional arguments}
}
\description{
generate network based on Enrichment results
}
\author{
Kai Guo
}
