% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{richR-internal}
\alias{richR-internal}
\alias{fast_factor}
\alias{hyper_bench_vector}
\alias{name_table}
\alias{reverseList}
\alias{sf}
\alias{uniq}
\alias{arrange.Annot}
\alias{arrange.GSEAResult}
\alias{arrange.richResult}
\alias{as.data.frame.Annot}
\alias{as.data.frame.GSEAResult}
\alias{as.data.frame.richResult}
\alias{filter.Annot}
\alias{filter.GSEAResult}
\alias{filter.richResult}
\alias{group_by.Annot}
\alias{group_by.GSEAResult}
\alias{group_by.richResult}
\alias{mutate.Annot}
\alias{mutate.GSEAResult}
\alias{mutate.richResult}
\alias{rename.Annot}
\alias{rename.GSEAResult}
\alias{rename.richResult}
\alias{select.Annot}
\alias{select.GSEAResult}
\alias{select.richResult}
\alias{slice.Annot}
\alias{slice.GSEAResult}
\alias{slice.richResult}
\alias{summarise.Annot}
\alias{summarise.GSEAResult}
\alias{summarise.richResult}
\title{Internal richR functions and methods}
\description{
Internal functions, Rcpp helpers, and dplyr S3 method dispatchers.
These are not intended to be called directly by the user.
}
\keyword{internal}
