% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{richResult-class}
\alias{richResult-class}
\alias{show,richResult-method}
\alias{plot,richResult-method}
\alias{summary,richResult-method}
\title{Class "richResult"
This class represents the result of enrichment analysis.}
\description{
Class "richResult"
This class represents the result of enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{enrichment analysis results}

\item{\code{detail}}{genes included in significant terms and original information}

\item{\code{pvalueCutoff}}{cutoff pvalue}

\item{\code{pAdjustMethod}}{pvalue adjust method}

\item{\code{padjCutoff}}{pvalue adjust cutoff value}

\item{\code{genenumber}}{number of input genes}

\item{\code{organism}}{organism used}

\item{\code{ontology}}{biological ontology}

\item{\code{gene}}{Gene IDs}

\item{\code{keytype}}{Gene ID type}

\item{\code{sep}}{character string used to separate the genes when concatenating}
}}

\author{
Kai Guo
}
\keyword{classes}
