% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richScatter}
\alias{richScatter}
\alias{richScatter,richResult-method}
\alias{richScatter,data.frame-method}
\alias{richScatter,GSEAResult-method}
\alias{ggscatter}
\alias{ggScatter}
\title{Labeled scatter plot of enrichment results}
\usage{
richScatter(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richScatter}{richResult}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richScatter}{data.frame}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richScatter}{GSEAResult}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richScatter(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richScatter(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "#fee0d2",
  high = "#b2182b",
  point.size = 5,
  label.size = 3,
  short = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult, data.frame, or GSEAResult}

\item{top}{number of terms (default 20)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{usePadj}{colour by adjusted p-value?}

\item{low}{low-end gradient colour}

\item{high}{high-end gradient colour}

\item{point.size}{point size}

\item{label.size}{label size}

\item{short}{shorten term names}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Labeled scatter plot of enrichment results
}
\examples{
\dontrun{
  res <- richKEGG(gene, kodata = hsako)
  richScatter(res)
}
}
