% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richTermSim}
\alias{richTermSim}
\alias{richTermSim,richResult-method}
\alias{richTermSim,data.frame-method}
\alias{richTermSim,GSEAResult-method}
\alias{ggtermsim}
\alias{ggTermSim}
\title{Term similarity scatter plot (MDS)}
\usage{
richTermSim(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richTermSim}{richResult}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richTermSim}{data.frame}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richTermSim}{GSEAResult}(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richTermSim(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richTermSim(
  object,
  top = 20,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "grey80",
  high = "#b2182b",
  label.size = 3,
  short = FALSE,
  sep = ",",
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{richResult, data.frame, or GSEAResult}

\item{top}{number of terms (default 20)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{usePadj}{colour by adjusted p-value?}

\item{low}{low-end gradient colour}

\item{high}{high-end gradient colour}

\item{label.size}{label size}

\item{short}{shorten term names}

\item{sep}{gene-ID separator}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Term similarity scatter plot (MDS)
}
\examples{
\dontrun{
  res <- richKEGG(gene, kodata = hsako)
  richTermSim(res)
}
}
