% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_new_methods.R, R/zzz_ggplots_aliases.R
\name{richVolcano}
\alias{richVolcano}
\alias{richVolcano,GSEAResult-method}
\alias{richVolcano,data.frame-method}
\alias{ggvolcano}
\alias{ggVolcano}
\title{Volcano plot of GSEA gene sets}
\usage{
richVolcano(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  sig.color = "#b2182b",
  ns.color = "grey70",
  point.size = 3,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richVolcano}{GSEAResult}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  sig.color = "#b2182b",
  ns.color = "grey70",
  point.size = 3,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

\S4method{richVolcano}{data.frame}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  sig.color = "#b2182b",
  ns.color = "grey70",
  point.size = 3,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richVolcano(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  sig.color = "#b2182b",
  ns.color = "grey70",
  point.size = 3,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)

richVolcano(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  sig.color = "#b2182b",
  ns.color = "grey70",
  point.size = 3,
  filename = NULL,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{object}{GSEAResult or data.frame with GSEA results}

\item{top}{max terms (default 50)}

\item{pvalue}{p-value cutoff}

\item{padj}{adjusted p-value cutoff}

\item{sig.color}{colour for significant points}

\item{ns.color}{colour for non-significant points}

\item{point.size}{point size}

\item{filename}{save path}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Volcano plot of GSEA gene sets
}
\examples{
\dontrun{
  res <- richGSEA(genelist, gseadata)
  richVolcano(res)
}
}
