% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{s4-accessors}
\alias{s4-accessors}
\alias{row.names,Annot-method}
\alias{row.names,richResult-method}
\alias{row.names,GSEAResult-method}
\alias{rownames,Annot-method}
\alias{rownames,richResult-method}
\alias{rownames,GSEAResult-method}
\alias{names,Annot-method}
\alias{names,richResult-method}
\alias{names,GSEAResult-method}
\alias{colnames,Annot-method}
\alias{colnames,richResult-method}
\alias{colnames,GSEAResult-method}
\title{S4 row- and column-related methods for Annot, richResult, GSEAResult}
\usage{
\S4method{row.names}{Annot}(x)

\S4method{row.names}{richResult}(x)

\S4method{row.names}{GSEAResult}(x)

\S4method{rownames}{Annot}(x)

\S4method{rownames}{richResult}(x)

\S4method{rownames}{GSEAResult}(x)

\S4method{names}{Annot}(x)

\S4method{names}{richResult}(x)

\S4method{names}{GSEAResult}(x)

\S4method{colnames}{Annot}(x)

\S4method{colnames}{richResult}(x)

\S4method{colnames}{GSEAResult}(x)
}
\arguments{
\item{x}{An object of class Annot, richResult, or GSEAResult}

\item{...}{Further arguments (ignored)}
}
\description{
These methods let you get rownames, colnames, names, etc. for your custom S4 classes.
}
