% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{searchPathways}
\alias{searchPathways}
\title{Search pathways by keyword}
\usage{
searchPathways(gseaRes, keyword, ignore_case = TRUE)
}
\arguments{
\item{gseaRes}{GSEAResult object}

\item{keyword}{search term (can be regex pattern)}

\item{ignore_case}{whether to ignore case when searching}
}
\description{
Search pathways by keyword
}
\examples{
\dontrun{
signaling_pathways <- searchPathways(gseaRes, "signaling")
mapk_pathways <- searchPathways(gseaRes, "MAPK")
}
}
\author{
Kai Guo
}
