% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{show-methods}
\alias{show-methods}
\alias{show,richResult-method}
\alias{show,GSEAResult-method}
\alias{show,Annot-method}
\title{Show methods for richR S4 classes}
\usage{
\S4method{show}{richResult}(object)

\S4method{show}{GSEAResult}(object)

\S4method{show}{Annot}(object)
}
\arguments{
\item{object}{An S4 object of class richResult, GSEAResult, or Annot}
}
\description{
Display a concise summary when an object is printed at the console.
}
