% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/species.R
\name{showData}
\alias{showData}
\title{show available data based on bioconductor annotation package}
\usage{
showData()

showData()
}
\value{
data.frame with dbname and species columns
}
\description{
show available data based on bioconductor annotation package

Show available species and their OrgDb packages
}
\author{
Kai Guo
}
