% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset-methods}
\alias{subset-methods}
\alias{[,Annot,ANY,ANY,ANY-method}
\alias{[,Annot-method}
\alias{[,richResult,ANY,ANY,ANY-method}
\alias{[,richResult-method}
\alias{[,GSEAResult,ANY,ANY,ANY-method}
\alias{[,GSEAResult-method}
\title{S4 subsetting methods for Annot, richResult, and GSEAResult}
\usage{
\S4method{[}{Annot,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{richResult,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{GSEAResult,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An S4 object (Annot, richResult, or GSEAResult)}

\item{i, j}{indices for subsetting}

\item{...}{further arguments (usually ignored)}

\item{drop}{logical. Should dimensions be dropped?}
}
\description{
These methods define `[` subsetting on each respective S4 class.
}
