% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{summary,GSEAResult-method}
\alias{summary,GSEAResult-method}
\title{S4 summary method for GSEAResult objects}
\usage{
\S4method{summary}{GSEAResult}(object, ...)
}
\arguments{
\item{object}{GSEAResult object}

\item{...}{Not used}
}
\value{
Prints summary to console
}
\description{
S4 summary method for GSEAResult objects
}
