% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{summary,richResult-method}
\alias{summary,richResult-method}
\title{S4 summary method for richResult objects}
\usage{
\S4method{summary}{richResult}(object, ...)
}
\arguments{
\item{object}{richResult object}

\item{...}{Not used}
}
\value{
Prints summary to console
}
\description{
S4 summary method for richResult objects
}
