% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{vec_to_df}
\alias{vec_to_df}
\title{vector to data.frame}
\usage{
vec_to_df(x, name)

vec_to_df(x, name)
}
\arguments{
\item{x}{named vector}

\item{name}{character vector of length 2 for column names}
}
\value{
data.frame
}
\description{
vector to data.frame

Convert a named vector to a two-column data.frame
}
\keyword{internal}
