% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.clean.char}
\alias{.clean.char}
\title{Remove newlines}
\usage{
.clean.char(x)

.clean.char(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector without newlines.
}
\description{
Replaces any newline characters with a space.
}
\examples{
.clean.char("First line\nSecond line\nThird line")
}
