% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{enrich}
\alias{enrich}
\title{enrich method}
\usage{
enrich(x, annot, pvalue = 0.05, ...)
}
\arguments{
\item{x}{vector contains gene names}

\item{annot}{Annotation object or dataframe}

\item{...}{additional parameters}
}
\value{
richResult
}
\description{
enrich
}
\author{
Kai Guo
}
