% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggbar}
\alias{ggbar}
\title{ggbar method}
\usage{
ggbar(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  ...
)
}
\arguments{
\item{object}{richResult or dataframe}

\item{...}{additional parameters}
}
\value{
plot
}
\description{
ggbar
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggbar(res)
}
}
\author{
Kai Guo
}
