% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.R
\name{ggbar_internal}
\alias{ggbar_internal}
\alias{ggbar,richResult-method}
\alias{ggbar,data.frame-method}
\title{barplot for Enrichment results}
\usage{
ggbar_internal(
  resultFis,
  top = 50,
  pvalue = 0.05,
  order = FALSE,
  horiz = TRUE,
  low = "lightpink",
  high = "red",
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  fontsize.text = 3,
  angle = 75,
  padj = NULL,
  usePadj = TRUE,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8
)

\S4method{ggbar}{richResult}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  fontsize.text = 3,
  angle = 75,
  orderp = orderp,
  filename = NULL,
  width = 10,
  height = 8,
  horiz = TRUE,
  ...
)

\S4method{ggbar}{data.frame}(
  object,
  top = 50,
  pvalue = 0.05,
  padj = NULL,
  order = FALSE,
  usePadj = TRUE,
  fontsize.x = 10,
  fontsize.y = 10,
  short = FALSE,
  fontsize.text = 3,
  angle = 75,
  orderp = FALSE,
  filename = NULL,
  width = 10,
  height = 8,
  horiz = TRUE,
  ...
)
}
\arguments{
\item{top}{number of terms you want to display,}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{order}{order by Term or richFactor}

\item{horiz}{horiz or not}

\item{low}{low color}

\item{high}{high color}

\item{font.x}{font of x axis}

\item{fontsize.x}{fontsize of x axis}

\item{fontsize.y}{fontsize of y axis}

\item{short}{automatic short name or not}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use p adjust value as color or not (should use with padj)}

\item{orderp}{order by p value(adjusted p value)}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{object}{dataframe of enrichment results}

\item{alpha}{transparency alpha}

\item{font}{y font of y axis}

\item{font.size}{font size for xlim or ylim}
}
\description{
barplot for Enrichment results

barplot for Enrichment result
}
\examples{
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggbar(res)
}
\dontrun{
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  gene=sample(unique(hsako$GeneID),1000)
  res<-richKEGG(gene,kodata = hsako)
  ggbar(result(res))
}
}
\author{
Kai Guo
}
