% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{ggGSEA}
\alias{ggGSEA}
\title{plot the gsea result}
\usage{
ggGSEA(x, term, object, gseaRes = gseaRes)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{term}{the significant term}

\item{object}{Annot object}

\item{gseaRes}{GSEAResult object}

\item{info}{Term with annotation details}
}
\description{
plot the gsea result
}
\author{
Kai Guo
}
