% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetmap.R
\name{ggnetmap}
\alias{ggnetmap}
\title{generate network based on Enrichment results}
\usage{
ggnetmap(
  richRes,
  gene = NULL,
  top = 50,
  top.display = NULL,
  pvalue = 0.05,
  padj = NULL,
  usePadj = TRUE,
  low = "orange",
  high = "red",
  weightcut = 0.2,
  useTerm = TRUE,
  writeCyt = FALSE,
  cytoscapeFile = "cytoscape.txt",
  cytoscapeFormat = "graphml",
  label.color = "black",
  label.size = 2,
  node.shape = NULL,
  layout = "fruchtermanreingold",
  savefig = FALSE,
  visNet = FALSE,
  smooth = TRUE,
  nodeselect = FALSE,
  edit = FALSE,
  savehtml = FALSE,
  filename = "network",
  width = 7,
  height = 7,
  segment.size = 0.2,
  node.alpha = 0.7,
  ...
)
}
\arguments{
\item{richRes}{list of enrichment object}

\item{gene}{vector contains gene names or dataframe with DEGs information}

\item{top}{number of terms to display}

\item{top.display}{top number to display}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}

\item{usePadj}{use adjust p value as color or not (should use with padj)}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{weightcut}{cutoff valule for edge}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{cytoscapeFormat}{Character string giving the output file format}

\item{label.color}{label color}

\item{label.size}{label size}

\item{layout}{layout method ('fruchtermanreingold','kamadakawai','target','circle')}

\item{savefig}{save the figure or not}

\item{filename}{figure output name}

\item{width}{figure width}

\item{height}{figure height}

\item{segment.size}{size for label segment}

\item{node.alpha}{alpha-transparency scales}

\item{label.font}{label font}
}
\description{
generate network based on Enrichment results
}
\examples{
\dontrun{
hsako <- buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
hsago <- buildAnnot(species="human",keytype="SYMBOL",anntype = "GO")
gene <- sample(unique(hsako$GeneID),1000)
resko <-richKEGG(gene,kodata = hsako)
resgo <- richGO(gene,hsago)
ggnetmap(list(resgo,resko))
}
}
\author{
Kai Guo
}
