% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{GSEAResult-class}
\alias{GSEAResult-class}
\alias{show,GSEAResult-method}
\alias{summary,GSEAResult-method}
\alias{plot,GSEAResult-method}
\title{Class "GSEAResult"
This class represents the result from GSEA analysis}
\description{
Class "GSEAResult"
This class represents the result from GSEA analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{result from fgsea analysis}

\item{\code{sig}}{genes included in significant terms and original information}

\item{\code{organism}}{organism}

\item{\code{keytype}}{Gene ID type}

\item{\code{params}}{parameters used for GSEA analysis}
}}

\author{
Kai Guo
}
\keyword{classes}
