% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{plot multiple significant pathways with enhanced customization and directional coloring}
\usage{
plotGSEA(
  object,
  gseaRes,
  mycol = NULL,
  up_color = "#E31A1C",
  down_color = "#1F78B4",
  show_direction = TRUE,
  pathways = NULL,
  pathway_pattern = NULL,
  category = NULL,
  min_genes = NULL,
  max_genes = NULL,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  gseaParam = 1,
  ticksSize = 0.2,
  plot_title = "GSEA Enrichment Plot",
  show_pathway_names = FALSE,
  pathway_name_size = 10,
  line_alpha = 0.8,
  point_alpha = 0.6,
  legend_position = "bottom",
  legend_ncol = 2,
  return_data = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{object}{Annot object}

\item{gseaRes}{GSEAResult object}

\item{mycol}{a vector indicate the colors used for the figure (used when show_direction = FALSE)}

\item{up_color}{color for up-regulated pathways (positive NES)}

\item{down_color}{color for down-regulated pathways (negative NES)}

\item{show_direction}{logical, whether to use directional coloring based on NES}

\item{pathways}{specific pathway names to display (if NULL, use top/pvalue/padj filtering)}

\item{pathway_pattern}{regex pattern for pathway matching}

\item{category}{pathway category filter}

\item{min_genes}{minimum gene set size filter}

\item{max_genes}{maximum gene set size filter}

\item{top}{number of terms you want to display (used when pathways is NULL)}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL, used when pathways is NULL)}

\item{padj}{cutoff value of p adjust value (used when pathways is NULL)}

\item{gseaParam}{GSEA parameter for enrichment score calculation}

\item{ticksSize}{size of the tick marks}

\item{plot_title}{title for the plot}

\item{show_pathway_names}{whether to show pathway names on plot}

\item{pathway_name_size}{size of pathway name text}

\item{line_alpha}{transparency of lines}

\item{point_alpha}{transparency of points}

\item{legend_position}{position of legend ("top", "bottom", "left", "right", "none")}

\item{legend_ncol}{number of columns in legend}

\item{return_data}{whether to return plot data along with plot}

\item{interactive}{whether to create interactive plot using plotly}
}
\description{
plot multiple significant pathways with enhanced customization and directional coloring
}
\examples{
\dontrun{
set.seed(123)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
# Plot with directional coloring (default)
plotGSEA(object = hsako, gseaRes = res, show_direction = TRUE)
# Plot specific pathways
plotGSEA(object = hsako, gseaRes = res,
         pathways = c("MAPK signaling pathway", "PI3K-Akt signaling pathway"))
# Plot pathways matching pattern
plotGSEA(object = hsako, gseaRes = res, pathway_pattern = "signaling")
# Traditional multi-color approach
plotGSEA(object = hsako, gseaRes = res, show_direction = FALSE)
}
}
\author{
Kai Guo
}
