% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{plot multiple significant pathways}
\usage{
plotGSEA(
  x,
  object,
  gseaRes,
  mycol = NULL,
  top = 10,
  pvalue = 0.05,
  padj = NULL,
  gseaParam = 1,
  ticksSize = 0.2
)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{object}{Annot object}

\item{gseaRes}{GSEAResult object}

\item{mycol}{a vector indicate the colors used for the figure}

\item{top}{number of terms you want to display,}

\item{pvalue}{cutoff value of pvalue (if padj set as NULL)}

\item{padj}{cutoff value of p adjust value}
}
\description{
plot multiple significant pathways
}
\examples{
\dontrun{
set.seed(123)
hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
name=sample(unique(hsako$GeneID),1000)
gene<-rnorm(1000)
names(gene)<-name
res<-richGSEA(gene,object = hsako)
plotGSEA(hsako,res,object = hsako)
}
}
\author{
Kai Guo
}
