% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGO.R
\name{richGO,ANY,Annot-method}
\alias{richGO,ANY,Annot-method}
\title{GO Enrichment analysis function}
\usage{
\S4method{richGO}{ANY,Annot}(
  x,
  godata,
  ontology = "BP",
  pvalue = 0.05,
  padj = NULL,
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  filename = NULL,
  padj.method = "BH"
)
}
\arguments{
\item{x}{vector contains gene names or dataframe with DEGs information}

\item{godata}{GO annotation data}

\item{ontology}{BP,MF or CC}

\item{pvalue}{cutoff pvalue}

\item{padj}{cutoff p adjust value}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{organism}{organism}

\item{keytype}{keytype for input genes}
}
\description{
GO Enrichment analysis function
}
\author{
Kai Guo
}
