% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{richGSEA}
\alias{richGSEA}
\title{richGSEA method}
\usage{
richGSEA(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  minSize = 15,
  ontology = "",
  maxSize = 500,
  nperm = 5000,
  filename = NULL,
  padj.method = "BH",
  organism = NULL,
  sep = ",",
  ...
)
}
\arguments{
\item{x}{vector contains gene names}

\item{object}{Annot object or dataframe}

\item{...}{additional parameters}
}
\value{
GSEAResult
}
\description{
richGSEA
}
\examples{
\dontrun{
set.seed(123)
  hsako<-buildAnnot(species="human",keytype="SYMBOL",anntype = "KEGG")
  name=sample(unique(hsako$GeneID),1000)
  gene<-rnorm(1000)
  names(gene)<-name
  res<-richGSEA(gene,kodata = hsako)
}
}
\author{
Kai Guo
}
