% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richGSEA.R
\name{richGSEA_internal}
\alias{richGSEA_internal}
\title{Enrichment analysis for any type of annotation data}
\usage{
richGSEA_internal(
  x,
  object,
  keytype = "",
  pvalue = 0.05,
  padj = NULL,
  minSize = 15,
  maxSize = 500,
  nperm = 5000,
  filename = NULL,
  padj.method = "BH",
  organism = NULL,
  ontology = NULL,
  table = TRUE
)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{object}{annotation file for all genes}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{filename}{output filename}

\item{padj.method}{p value adjust method}

\item{table}{leadingEdge as vector}
}
\description{
Enrichment analysis for any type of annotation data
}
\author{
Kai Guo
}
