% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary,GSEAResult-method}
\alias{summary,GSEAResult-method}
\title{S4 summary method for GSEAResult objects}
\usage{
\S4method{summary}{GSEAResult}(object, ...)
}
\arguments{
\item{object}{A \code{GSEAResult} object}

\item{...}{Not used}
}
\value{
Prints summary info to the console
}
\description{
Summarizes the content of a \code{GSEAResult} object.
}
