% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary,richResult-method}
\alias{summary,richResult-method}
\title{S4 summary method for richResult objects}
\usage{
\S4method{summary}{richResult}(object, ...)
}
\arguments{
\item{object}{A \code{richResult} object}

\item{...}{Not used}
}
\value{
Prints summary info to the console
}
\description{
Summarizes the content of a \code{richResult} object.
}
